Add vscode debugging settings
